GUI Version 1.2.3.11 [DEV 2.1]
Changes:
- Added checking for updates (when launching the application)
- Minor edit
