+++ sources/GUI/Data.cs
@@ -1,4 +1,5 @@
 ﻿static class Data
-    {
+{
+        public static string JetpackDowngraderVersion { get; } = "2.3";
         public static string PathToGame { get; set; }
     }
